+++ packages/helpers/CHANGELOG.md
@@ -1,3 +1,7 @@
+
+- Support `wrapUri()` method
+- Fix: ExplodeVeridaUri regex breaking when there are no query params
+
 2023-02-16 (v2.1.1)
 -------------------